Version 1.7
StmtNode Has been written and commented, but not tested.
StmtTailNode Has been written and commented, but not tested.
IfStmtNode Has been written and commented, but not tested.
LoopStmtNode Has been written and commented, but not tested.
BreakStmtNode Has been written and commented, but not tested.
AssignStmtNode Has been written and commented, but not tested.
ReadStmtNode has been written and commented, but not tested.
PrintStmtNode has been written and commented, but not tested